changed searchbar to material 3 and added clear text btn
