+++ app/src/main/java/com/example/dictionaryapp/ui/components/DisplayMeaning.kt
@@ -1,24 +1,9 @@
 package com.example.dictionaryapp.ui.components
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import com.example.dictionaryapp.ui.screens.DictionaryUiState
-import androidx.compose.material.icons.filled.ErrorOutline
-import androidx.compose.material3.CircularProgressIndicator
+import com.example.dictionaryapp.ui.screens.ErrorScreen
+import com.example.dictionaryapp.ui.screens.LoadingScreen
 
 @Composable
 fun DisplayMeaning(dictionaryUiState: DictionaryUiState) {
@@ -26,51 +11,5 @@ fun DisplayMeaning(dictionaryUiState: DictionaryUiState) {
         is DictionaryUiState.Error -> ErrorScreen(e = dictionaryUiState.e)
         is DictionaryUiState.Loading -> LoadingScreen()
         is DictionaryUiState.Success -> DisplayDefinition(definition = dictionaryUiState.definition)
-    }
-}
-
-@Composable
-fun ErrorScreen(e: Exception) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(24.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Icon(
-            imageVector = Icons.Default.ErrorOutline,
-            contentDescription = "Error",
-            tint = MaterialTheme.colorScheme.error,
-            modifier = Modifier.size(64.dp)
-        )
-        Spacer(modifier = Modifier.height(16.dp))
-        Text(
-            text = "Oops! Something went wrong.",
-            style = MaterialTheme.typography.titleMedium,
-            color = MaterialTheme.colorScheme.error
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        Text(
-            text = e.localizedMessage ?: "Unknown error",
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.padding(horizontal = 8.dp),
-            maxLines = 3
-        )
-    }
-}
-
-@Composable
-fun LoadingScreen() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        CircularProgressIndicator(
-            color = MaterialTheme.colorScheme.primary,
-            strokeWidth = 4.dp,
-            modifier = Modifier.size(48.dp)
-        )
     }
 }
